WA-271 Renamed and added in memory gatekeeper repository tests
